requirements.txt
tensorflow>=2.15.0,<3.0.0
numpy>=1.24.0
opencv-python>=4.8.0
scikit-learn>=1.0.0
streamlit>=1.28.0
streamlit-webrtc>=0.47.3
streamlit-notifications>=0.1.0
pyttsx3>=2.90
av>=10.0.0
pillow>=10.0.0
python-dotenv>=1.0.0
aiofiles>=23.1.0
